ai_service/requirements.txt
fastapi>=0.110
uvicorn>=0.27
chromadb>=0.4
openai>=1.0
python-dotenv>=1.0
tiktoken>=0.6